\name{example_luxA}
\alias{example_luxA}

\title{
Example Data for Protein Sequences
}

\description{
The data consist of selected amino acid sequences of the luxA. There are 2283 aligned sequences of length 360.
}

\usage{
example_luxA
}

\value{
aligned amino acid sequences
}

\source{
Hawkins-Hooker, A., Depardieu, F., Baur, S., Couairon, G., Chen, A., & Bikard, D. (2020). Generating functional protein variants with variational autoencoders. bioRxiv.
}

\author{
Dongmin Jung
}

