\name{example_PTEN}
\alias{example_PTEN}

\title{
Example Data for Protein Sequences
}

\description{
The data consist of selected amino acid sequences of the PTEN. There are 912 aligned sequences of length 403.
}

\usage{
example_PTEN
}

\value{
aligned amino acid sequences
}

\source{
Frazer, J., Notin, P., Dias, M., Gomez, A., Brock, K., Gal, Y., & Marks, D. (2020). Large-scale clinical interpretation of genetic variants using evolutionary data and deep learning. bioRxiv.

}

\author{
Dongmin Jung
}

